\name{emxVarianceComponents}
\alias{emxVarianceComponents}

\title{Creates Variance Components Model}

\description{
This function creates a variance components model as an MxModel object.
}

\usage{
emxVarianceComponents(model, data, run)
}


\arguments{
   \item{model}{MxModel. See Details.}
   \item{data}{matrix or data.frame. The used in the model}
   \item{run}{logical. Whether to run the model before returning.}
}

\details{
This function does not really do anything currently.  Do not use it.
}

\value{
    MxModel.
}

\seealso{
\link{emxFactorModel}
}

%\references{
%
%}

\examples{
   
# Create a loadings matrix
require(EasyMx)


}

